refactor(ProfileImage|CustomImage): Move to Surfaces
